Remove math and generic_math from lyon_geom.
