+++ tessellation/src/earcut_tests.rs
@@ -19,7 +19,7 @@
 // THIS SOFTWARE.
 // -------------
 
-use crate::geom::math::*;
+use crate::math::*;
 use crate::geometry_builder::{simple_builder, VertexBuffers};
 use crate::path::{Path, PathSlice};
 use crate::FillOptions;
