+++ tessellation/src/fill_tests.rs
@@ -1,5 +1,5 @@
 use crate::extra::rust_logo::build_logo_path;
-use crate::geom::math::*;
+use crate::math::*;
 use crate::geometry_builder::*;
 use crate::path::builder::{Build, PathBuilder};
 use crate::path::{Path, PathSlice};
